Fixed a bug parsing short values with RPN.
git-svn-id: https://svn.aps.anl.gov/AOP/oag/trunk/apps/src/elegant@24509 c1b61890-85f3-ea29-e4d6-a272fe79f168

+++ mad_parse.c
@@ -976,7 +976,7 @@ void parse_element(
       if (!isdigit(*ptr) && *ptr!='-' && *ptr!='+') {
         rpn_token = get_token(ptr);
         SDDS_UnescapeQuotes(rpn_token, '"');
-        *((long*)(p_elem+parameter[i].offset)) = rpn(rpn_token);
+        *((short*)(p_elem+parameter[i].offset)) = rpn(rpn_token);
         if (rpn_check_error()) exitElegant(1);
         printf("computed value for %s.%s is %hd\n", 
                 eptr->name, parameter[i].name,
